extract "openDocumentAndFind" function
